requirements.txt
Django>=5.0.0
Pillow>=10.0.0
qrcode>=7.4.0
faker
gunicorn>=21.2.0
whitenoise>=6.6.0